redesign teach point, multi-try-one-selected
redesign teach point method, calc the distance between
teach point and current weld point, use exp the decide
the effect ratio, this is a ratio to be multiplied to
the sigma of each axes.

on seam point 0, try multi times(80), select different
local optimum as the start values, plan a sample of
10 percent seam point, calc the global criteria average
as the recommend criteria, select the biggest recommend
criteria solution, then use this solution's start point
to plan the whole job.